+++ src/get_course_data.py
@@ -1,12 +1,9 @@
 import pandas as pd
 from canvasapi import Canvas
-import getpass
-import sys
 from IPython.display import display, HTML
 from datetime import datetime
-from interface import shut_down, print_success, _create_csv
-from helpers import create_folder
-
+from interface import shut_down, print_success
+import helpers
 import os
 from pathlib import Path
 from settings import COURSE_ID
@@ -38,64 +35,25 @@ def create_canvas_object():
         
     except Exception as e:
         shut_down(f'{e}: Canvas object not created')
-        return(False)
-        
- 
-def create_df_and_csv(paginatedlist, output_file, filter_to_columns=None, keep=True):
-    #TODO - figure out "best" structure for this kind of data
-    
-    """given a list of objects or paginatedlist return a dataframe
-    
-    Args:
-        paginatedlist (a Canvas PaginatedList)
-        output_file (str)
-        filter_to_columns (None or list)
-        keep (bool)
-    
-    Returns:
-        df (dataframe) 
-        
-    Output:
-        csv in output_path if data available
-        
-    """
-    
-    try:
-        df = pd.DataFrame([i.__dict__ for i in paginatedlist])
-        
-        # if includes filtered columns list then keep if keep=True
-        # or drop if keep=False
-        if filter_to_columns:
-            
-            if keep:
-                df = df[filter_to_columns]
-            else:
-                df.drop(filter_to_columns, axis=1, inplace=True)
-        
-        df.to_csv(f'{output_file}.csv')
-        
-        return(df)
-        
-    except Exception as e:
-        print(f'no dataframe for {output_file}: {e}')
+        return(False)       
 
 def get_course_data(course, output_path):
-    enrol_df = create_df_and_csv(course.get_enrollments(), f'{output_path}/enrollments')
-    files_df = create_df_and_csv(course.get_files(), f'{output_path}/files')
-    features_df = create_df_and_csv(course.get_features(), f'{output_path}/features')
-    pages_df = create_df_and_csv(course.get_pages(), f'{output_path}/pages')
-    quizzes_df = create_df_and_csv(course.get_quizzes(), f'{output_path}/quizzes')
-    assignments_df = create_df_and_csv(course.get_assignments(), f'{output_path}/assignments')
-    externaltools_df = create_df_and_csv(course.get_external_tools(), f'{output_path}/external_tools')
-    tabs_df = create_df_and_csv(course.get_tabs() , f'{output_path}/tabs')
-    discussion_topics_df = create_df_and_csv(course.get_discussion_topics(), f'{output_path}/discussion_topics')
-    grades_df = create_df_and_csv(course.get_multiple_submissions(student_ids='all'), f'{output_path}/assignment_submissions')
+    enrol_df = helpers.create_df_and_csv(course.get_enrollments(), f'{output_path}/enrollments')
+    files_df = helpers.create_df_and_csv(course.get_files(), f'{output_path}/files')
+    features_df = helpers.create_df_and_csv(course.get_features(), f'{output_path}/features')
+    pages_df = helpers.create_df_and_csv(course.get_pages(), f'{output_path}/pages')
+    quizzes_df = helpers.create_df_and_csv(course.get_quizzes(), f'{output_path}/quizzes')
+    assignments_df = helpers.create_df_and_csv(course.get_assignments(), f'{output_path}/assignments')
+    externaltools_df = helpers.create_df_and_csv(course.get_external_tools(), f'{output_path}/external_tools')
+    tabs_df = helpers.create_df_and_csv(course.get_tabs() , f'{output_path}/tabs')
+    discussion_topics_df = helpers.create_df_and_csv(course.get_discussion_topics(), f'{output_path}/discussion_topics')
+    grades_df = helpers.create_df_and_csv(course.get_multiple_submissions(student_ids='all'), f'{output_path}/assignment_submissions')
     
 
     # discussion topics, entries and replies
     #modules and module items
     modules = course.get_modules()
-    modules_df = create_df_and_csv(modules, f'{output_path}/modules')
+    modules_df = helpers.create_df_and_csv(modules, f'{output_path}/modules')
 
     module_items = []
 
@@ -121,9 +79,9 @@ def create_course_output():
     #create a project structure for the new course
     course = canvas.get_course(COURSE_ID)
     
-    create_folder(settings.APIOUTPUT_FOLDER)
-    create_folder(settings.NEWANALYTICS_FOLDER)
-    create_folder(settings.GRADEBOOK_FOLDER) 
+    helpers.create_folder(settings.APIOUTPUT_FOLDER)
+    helpers.create_folder(settings.NEWANALYTICS_FOLDER)
+    helpers.create_folder(settings.GRADEBOOK_FOLDER) 
 
     get_course_data(course, settings.APIOUTPUT_FOLDER)
